detection-service/pyproject.toml
[project]
name = "hidrogreen-detectionservice"
version = "0.0.1"
description = "Service to detect the desease of the plant using the model"
authors = [
    {name = "Diego22rct",email = "diegorafaelcisnerostafur@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
package-mode = false
dependencies = [
    "fastapi[standard] (>=0.115.12,<0.116.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "tensorflow (>=2.19.0,<3.0.0)",
    "pillow (>=11.2.1,<12.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

